+++ CHANGES.md
@@ -2,7 +2,11 @@
 
 ## Unreleased
 
+### Features
+- It is now possible to provide source paths in configuration file [#154](https://github.com/MarketSquare/robotframework-tidy/issues/154)
+
 ## 1.5.1
+
 ### Fixes
 - Robotidy will now not crash on directory path [#177](https://github.com/MarketSquare/robotframework-tidy/issues/177)